chalkboard/requirements.txt
django==1.7.4
django-registration-redux
django-tokenapi
